fix: resolve SonarCloud MAJOR issues across codebase
S7172: replace implicit optional bool casts with has_value()
S6023: use value_or() for optional with default
S5827: replace redundant declared types with auto
S5276: add explicit static_cast for int/size_t→float/double conversions
S5817: mark non-mutating member functions const
S5945: replace C arrays with std::array / std::string
S6003: push_back → emplace_back
S6030: emplace → try_emplace
S6186: remove redundant operator!= synthesized by C++20 operator==
S3562: add default case to switch
S824: move block-scope extern declaration to file scope
S1066: merge nested if statements
S1121: extract assignment from condition
S3230: add in-class initializer for keybindings member
S1144: remove unused xterm_color function

+++ libs/draxul-font/src/glyph_atlas_manager.h
@@ -39,7 +39,7 @@ public:
             return region;
 
         // Atlas overflowed — reset and retry once.
-        glyph_cache_.reset(resolver.primary().face(), resolver.primary().point_size());
+        glyph_cache_.reset(resolver.primary().face(), static_cast<int>(resolver.primary().point_size()));
         atlas_reset_pending_ = true;
         atlas_reset_count_++;
         DRAXUL_LOG_WARN(
